Sorting changes

+++ models/getArticles.model.js
@@ -4,7 +4,7 @@ const devTopics = require('../db/data/development-data/topics')
 
 function getAllArticles(topic, sort, order) {
     let queryString = `
-    SELECT articles.author, articles.title, articles.article_id, articles.topic, articles.created_at, articles.votes, articles.article_img_url, COUNT(comments.comment_id) AS comment_count
+    SELECT articles.author, articles.title, articles.article_id, articles.topic, articles.created_at, articles.votes, articles.article_img_url, COUNT(comments.comment_id)::integer AS comment_count
     FROM articles
     LEFT JOIN comments ON articles.article_id = comments.article_id`
